+++ app/Http/Controllers/TshirtImageController.php
@@ -110,22 +110,10 @@ class TshirtImageController extends Controller
 
     public function create(): View
     {   
-        if(Auth::user()->isCustomer()){
-
-                $categories = Category::all();
-                $tshirtImage = new TshirtImage();
-                $tshirtImage->category_id = null;
-                $tshirtImage->customer_id = Auth::user()->id;
-                $tshirtImage->image_url = "teste";
-
-            return view('tshirt_images.create', compact('tshirtImage' ,'categories'));
-        }elseif(Auth::user()->isAdmin()){
-
-                $categories = Category::all();
-                $tshirtImage = new TshirtImage();
-                $tshirtImage->customer_id = null;
-                $tshirtImage->image_url = "teste";
-            return view('tshirt_images.create', compact('tshirtImage' ,'categories'));
+        if(Auth::user()->isAdmin() || Auth::user()->isCustomer()){
+            $tshirtImage = new TshirtImage();
+            $categories = Category::all();
+            return view('tshirt_images.create' , compact('categories', 'tshirtImage'));
         }else{
             return view('home')->with('alert-msg', 'Não tem permissões para criar imagens de tshirts!')->with('alert-type', 'danger');
         }
@@ -134,9 +122,10 @@ class TshirtImageController extends Controller
 
     public function edit(TshirtImage $tshirtImage): View
     {
-        if(Auth::user()->isAdmin()){
+        if(Auth::user()->isAdmin() || Auth::user()->id == $tshirtImage->customer_id){
             $categories = Category::all();
-            return view('tshirt_images.edit', compact('tshirtImage', 'categories'));
+            $customer = Customer::where('id', $tshirtImage->customer_id)->first();
+            return view('tshirt_images.edit', compact('tshirtImage', 'categories', 'customer'));
         }else{
             return view('home')->with('alert-msg', 'Não tem permissões para editar imagens de tshirts!')->with('alert-type', 'danger');
         }
@@ -144,7 +133,7 @@ class TshirtImageController extends Controller
 
     public function update(Request $request, TshirtImage $tshirtImage): View
     {   
-        if(Auth::user()->isAdmin()){
+        if(Auth::user()->isAdmin() || Auth::user()->id == $tshirtImage->customer_id){
             $categories = Category::all();
             $tshirtImage->update($request->all());
             return view('tshirt_images.edit', compact('tshirtImage', 'categories'));
@@ -157,8 +146,8 @@ class TshirtImageController extends Controller
     {
         if(Auth::user()->isCustomer()) {
             $customer = $request->user()->customer;
-            $tshirtImages = $customer->tshirt_images;
-
+            $tshirtImages = $request->user()->customer->tshirt_images;
+        
             return view('tshirt_images.minhas', compact('tshirtImages', 'customer'));
         }else{
             return view('home')->with('alert-msg', 'Não é cliente, logo não tem imagens de tshirts!')->with('alert-type', 'danger');
@@ -167,27 +156,57 @@ class TshirtImageController extends Controller
 
     public function store(Request $request): RedirectResponse
     {
-        TshirtImage::create($request->all());
-        return redirect('/tshirt_images');
+        if(Auth::user()->isCustomer()){
+            $tshirtImage = new TshirtImage();
+            $tshirtImage->name = $request->input('name');
+            $tshirtImage->description = $request->input('description');
+            $tshirtImage->category_id = null;
+            $tshirtImage->customer_id = $request->user()->id;
+            $tshirtImage->image_url = "teste";
+            $tshirtImage->save();
+
+        return redirect()->route('tshirt_images.minhas')
+            ->with('alert-msg', 'Imagem criada com sucesso!')
+            ->with('alert-type', 'success');
+
+        }elseif(Auth::user()->isAdmin()){
+
+            $tshirtImage = new TshirtImage();
+            $tshirtImage->name = $request->input('name');
+            $tshirtImage->description = $request->input('description');
+            $tshirtImage->category_id = $request->input('category_id');
+            $tshirtImage->customer_id = null;
+            $tshirtImage->image_url = "teste";
+            $tshirtImage->save();
+
+            return redirect()->route('tshirt_images.index')
+                ->with('alert-msg', 'Imagem criada com sucesso!')
+                ->with('alert-type', 'success');
+        }else{
+            return redirect()->route('home')->with('alert-msg', 'Não tem permissões para criar imagens de tshirts!')->with('alert-type', 'danger');
+        }
     }
 
     public function destroy(TshirtImage $tshirt_image): RedirectResponse
     {
-        try {
-            $tshirt_image->delete();
-            if ($tshirt_image->image_url) {
-                Storage::delete('/storage/tshirt_images/' . $tshirt_image->image_url);
-            }
-            return redirect()->route('tshirt_images.index')
-                ->with('alert-msg', "Tshirt Image #{$tshirt_image->id} apagada com sucesso!")
-                ->with('alert-type', 'success');
-        } catch (\Exception $error) {
-            $url = route('tshirt_images.show', ['tshirt_image' => $tshirt_image]);
-            $htmlMessage = "Não foi possível apagar a Tshirt Image <a href='$url'>#{$tshirt_image->id}</a> porque ocorreu um erro!" . $error->getMessage();
-            $alertType = 'danger';
-            return back()
-                ->with('alert-msg', $htmlMessage)
-                ->with('alert-type', $alertType);
+        if(Auth::user()->isAdmin() || Auth::user()->id == $tshirt_image->customer_id)
+            try {
+                $tshirt_image->delete();
+                if ($tshirt_image->image_url) {
+                    Storage::delete('/storage/tshirt_images/' . $tshirt_image->image_url);
+                }
+                return redirect()->route('tshirt_images.index')
+                    ->with('alert-msg', "Tshirt Image #{$tshirt_image->id} apagada com sucesso!")
+                    ->with('alert-type', 'success');
+            } catch (\Exception $error) {
+                $url = route('tshirt_images.show', ['tshirt_image' => $tshirt_image]);
+                $htmlMessage = "Não foi possível apagar a Tshirt Image <a href='$url'>#{$tshirt_image->id}</a> porque ocorreu um erro!" . $error->getMessage();
+                $alertType = 'danger';
+                return back()
+                    ->with('alert-msg', $htmlMessage)
+                    ->with('alert-type', $alertType);
+        }else{
+            return redirect()->route('home')->with('alert-msg', 'Não tem permissões para apagar imagens de tshirts!')->with('alert-type', 'danger');
         }
     }
 }
